requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
djangorestframework==3.12.2
jwt==1.2.0
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
